M600: favor UNLOAD_FILAMENT instead of custom G-code and park in favorable location.
Instead of having to maintain a separate implementation of unloading filament, use the already established UNLOAD_FILAMENT macro.

Parking is also done "front and center" so that changing filament is an easier job. Center is used instead of the front right (like used on MK4) since the MK3 extruder can sometimes get a a bit jammed up when unloading and this way you have access to the left and right side of the toolhead in order to work on a potential jam.

The Z parameter and subsequent manual move was removed since its not possible to fully know the restoration context and assuming that the LOAD_FILAMENT then has to restore it is not safe.