Enable adaptive bed mesh calibration.
Native support for adaptive bed mesh calibration was added about a year ago and since this project already does object exclusion by default, this is an easy change to just pass ADAPTIVE=1 to the bed mesh calibration command.

Just make sure the label object setting is enabled in the slicer (it should be on by default).